throw an error when github: sources are used in the lockfile
We are currently unable to fetch github: sources in a variety of cases.
Often the sources do not properly specify the source branch (or tag) but
just refer to a short-rev that is then expanded into a "long" rev in the
lockfile. To use `builtins.fetchGit` we must know the proper ref (e.g. a
branch) to fetch that revision from.